+++ front/src/views/components/Oina/Oina.tsx
